para PID?

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Config/LiftUtil.java
@@ -11,7 +11,7 @@ public class  LiftUtil {
     public static double AutoVertSlidesError = 0;
 
 
-    public static double target = 12400
+    public static int target = 12400
             ;
     public static double downTarget = 0;
